backend/requirements.txt
fastapi==0.110.0          # Updated from 0.104.1
uvicorn==0.27.1          # Updated from 0.24.0
python-dotenv==1.0.1     # Updated from 1.0.0
pydantic==2.6.3          # Updated from 2.4.2
pydantic-settings==2.2.1 # Updated from 2.0.3
openai==1.12.0           # Updated from 1.3.0
setuptools>=65.5.1        # Added for pkg_resources
manim==0.17.3            # Downgraded to ensure compatibility
numpy>=1.19
scipy>=1.7.3
pillow>=9.1
pycairo>=1.21.0
sqlalchemy==2.0.27       # Updated from 2.0.23
psycopg2-binary==2.9.9   # Same version
python-jose==3.3.0       # Same version
passlib==1.7.4           # Same version
pytest==8.0.2            # Updated from 7.4.3
httpx==0.27.0            # Updated from 0.25.1
alembic==1.13.1          # Same version
redis==5.0.1             # Same version
aiofiles==23.2.1         # Same version
python-magic==0.4.27  # For file type detection